Save works and gui functionality added

+++ src/main/java/application/Item.java
@@ -2,7 +2,7 @@ package main.java.application;
 
 public class Item{
 
-    public Item(String aName, float aPrice){
+    public Item(String aName, double aPrice){
         this.name = aName;
         this.price = aPrice;
     }
